Move scheduler into its own directory and namespace.
The scheduler is only used by the runner at this time, but has no dependency
on the runner itself. This patch moves the scheduler and its associated
coroutine library (which previously lived in src/runner) into a
src/scheduler directory.

The Scheduler and Job class have been moved from the runner namespace to
the scheduler one. References in the runner, server and connection have
been updated to reflect this new organization.