Do not yield() on side-effect-free jobs
When a job declares that it has no side effect, it means that we are
probably evaluating a condition, which is expected to take a short time.
In this case, a job simply doing yield() will not be queued away. It will
however gets properly queued when it terminates or when it does a
yield_until_things_changed() and friends.

This cut the number of cycles by roughly 20% in my tests.

	* src/scheduler/scheduler.cc (resume_scheduler): Do not queue
	away a job which is not terminated and is side-effect-free.