Handle tricky cases where job can be killed in bizarre states.
A job can inherit tags at creation time. For example, in

  a: { f(x) & g(x) }

g(x) will start a runner with tag "a" set at creation time. As a
consequence, if "a.block" arrives while "g(x)" is running, the
runner will not be able to catch the BlockedException because it
has not been set under the runner responsibility.

Also, it was possible that the job was killed even before it
started. In this case, the exception was not caught properly in
the job methods as exceptions were supervised only during the
work() phase, which happens once the job has really started.

As a consequence, the following statement

  a: { 1 & a.block }

would fail and propagate BlockedException up to the job coroutine
creation point, making the program fail with a fatal error.

	 * src/scheduler/job.cc (run): Put the initial yield() under exception
	 handler supervision. Also, catch BlockedException at the top-level.
	 As an extra precaution, remove the link to other processes before
	 signalling them to avoid a potential (which can never happen
	 with the current code anyway) loop.
	 (check_for_pending_exception): Check for the blocked state early,
	 and do not propagate a late exception as it would make little
	 sense once the context has been popped.
	 * src/scheduler/scheduler.cc (execute_round): Make it clearer that
	 blocked jobs will be started unconditionally so that they can
	 pop their context and requeue themselves in the correct state.
	 (check_for_stopped_tags): No need to set the job state as running,
	 as it will be started unconditionally by "execute_round".
	 src/scheduler/scheduler.hh: Add comment.