Add metaprogramming utilities and visitor library.
I wrote this code on my free time for a few other projects. I'm
putting it directly in libport for now, thus entrusting it to you all:
it is (will be) GPL, and I grant full, potentially commercial
read/write access to it to Gostai. Please don't redistribute it
without telling me - and don't consider it's Gostai's property, JC
;-).

It includes several pieces of general-purpose metaprograming
libraries. We might want to move it to a separated repository since
there are a few other metaprograming stuff we may want to use.

The file 'tests/libport/visitor.cc' is also a quick start guide for
visitors.  If you see errors looking like:

Use of incomplete type Error<Hey_this_is_very_wrong<something, ...> >

It's a meta error emitted by the library, the message being the
argument to Error.

Note: the library needs cleanup. This is just sample code so as some
person that requested it can toy around with it.

    * include/libport/errors.hh: New. Metaprogramming error libraries.
    * include/libport/meta.hh: New. Base metaprogramming utilities.
    * include/libport/traits.hh: New. Traits collection. This is partly already
    implemented in select-const & cie.
    * include/libport/typelist.hh: New. Type list library.
    * include/libport/visitor.hh: New. Visitor library.
    * tests/libport/visitor.cc: New. Visitor test and example.
    * tests/libport/Makefile.am: Register the new test. Remove some that does
    not even compile anymore, since every one pushes in libport without running
    and/or updating the test suite :-(.