Use only refcounted jobs.
Now that we have intrusive shared pointers, we can work with
refcounted jobs more easily. It also means that we can get rid of the
self-reference which is handed to the scheduler when the job dies, as
the scheduler will keep a refcounted vector list of every existing
jobs in the current round. The job will then be destroyed the next
time the scheduler goes through the list by not being requeued.

	* src/scheduler/job.hh, src/scheduler/job.hxx,
	src/scheduler/job.cc (myself_, myself_get()): Remove.
	* src/runner/interpreter.cc, src/object/system-class.cc,
	src/object/task-class.cc: "myself_get" has been removed.
	* src/scheduler/scheduler.hh, src/scheduler/scheduler.hxx,
	* src/scheduler/scheduler.hh (take_job_reference): Remove.
	(jobs_get): Return a copy of the jobs list.
	Work with reference-counted jobs everywhere.