By default, make libport::shared_ptr intrusive.
We can't use the meta::Inherit magic here, since it answers 'No' if the
refcounted type is only forward declared. In this case, we could sometime
refcount classically some libport::refcounted classes, resulting in a multiple
free fiasco.

By default, libport::shared_ptr are intrusive - that is, the second template
parameter is true. This way, if one forget to specify it while it had to be
false, it will fail at compile time the refcounted object doesn't inherit from
libport::RefCounted. If the default were false, and one forgot to specify true
for a libport::refcounted object, it would compile and entail runtime errors
(multiple frees).

    * include/libport/shared-ptr.hh: Here.