Enable to add specific actions in pod_cast. Add some for smart pointers.
Specializing SpecialActions enables to add specific actions on every read/write
to the pod_cast, depending on the stored type.

Add an action to increment the reference counter of smart pointers when they're
stored into a pod_cast. Otherwise, they would delete their target when the last
reference is hidden in a pod_cast.

We could/should decrement the counter on read, but this impose that users read
once and only once every smart pointer pod_cast, which the parser doesn't do
(multiple reads). For now, it can be left as-is, the consequence being leaks.

    * include/libport/pod-cast.hh,
    * include/libport/pod-cast.hxx: There.