Simplify SingletonPtr.
This is simpler, and I can't see any difference with the previous code.
It is also much simpler to use, we don't need all these macros.  But I
wait for the merge of 1.5.1 to happen before simplifying the uses.

This patch seems to revert 32ec8812d4da0866788b74ca58db9a7963de9c9d,
but unfortunately there are two modifications in that patch: instance()
is made static, and the variable becomes a member.  I believe the
first change is enough.  Unfortunately there is no reference to
information about this, so I can only guess.

If you need to revert this patch, document the file!!!
	* include/libport/singleton-ptr.hh (instance): Use a static
	variable instead.