Start a new job less than a full cycle later after it has been created.
With the actual scheme, a job is started in the next cycle. The first
time it is launched, it doesn't do anything useful and starts working
for real one cycle later.

This is problematic in some cases where we want to catch some conditions
change as soon as they occur, and no later than one cycle after they
occur. For example, the following code:

function f(x)
{
  at (x == 2) return 2;
  x = x + 1;
  x * 2
}

would not trigger when called with f(1), while we clearly intend it to
do.

This change ensures that the new job (created by the "at jobs handler")
will start during the current cycle.

	* src/object/list-class.cc (each_and): Even if we do not
	guarantee the order, minimize the surprise by creating the jobs
	in the reverse order (they will be started in the regular order
	then).
	* src/runner/interpreter.cc: Do not call "reserve" on
	"jobs_type".
	* src/scheduler/fwd.hh: "jobs_type" is now a list as we need to
	insert right in the middle.
	* src/scheduler/job.cc (work): Do not "yield" at the beginning,
	we have some work to do. Also, set the job as "running" as soon
	as it is started to avoid confusing the scheduler if the job
	must be stopped right away. And do not forget to check for a
	pending exception before working.
	* src/scheduler/job.hh: Update comment.
	* src/scheduler/scheduler.cc (work): Use a list iterator.
	(add_job): Add the job at the right place.
	(signal_stop): Remove any non-started job from the starting
	blocks.
	* src/scheduler/scheduler.hh (job_p_): Make the iterator a field
	of the scheduler to be able to access it from "add_job".