sdk/local.mk instead of sdk/Makefile.am.
The first concern of this patch is correctness.  In the kernel for
instance, sdk/config.h is the master from which libport/config.h is
built.  Because of the order in SUBDIRS (chosen for other reasons),
libport was firt processed, and then sdk/.  As a result two "make" in a
row gave two builds: the first one because sdk/config.h was updated, and
then another one because of libport/config.h, finally updated.

No simple solution was found with several Makefiles, but with a single
one, Make can finally do its job: track and honor dependencies.

	* sdk/Makefile.am: Rename as...
	* sdk/local.mk: this.
	Adjust for inclusion instead of being stand alone Makefile.am.
	* build-aux/init.mk (nodist_bin_SCRIPTS, dist_bin_SCRIPTS):
	Initialize.
	* sdk/Makefile.am: New, in case someone insists on using it.o
	* sdk/urbi-sdk.m4: No longer create sdk/Makefile.