Workaround for libcoro 'tweak' not working under wine.
The code was expecting GetCurrentFiber to return a magic value when the
current thread is not a fiber. I found no trace of this voodoo stuff in
microsoft doc. However the clean API call to know if the current thread
is a fiber is only available for windows vista and up. So keep it and
patch it for wine which uses a different magic value (of 0).

	* src/scheduler/libcoroutine/Coro.cc(Coro_initializeMainCoro):
	  Also consider that current thread is not a fiber if
	  GetCurrentFiber() returns 0.
	  (Coro_switchTo_): 'Use' 'self' argument in USE_FIBERS case to
	  avoid a warning.