Add a "--enable-static-boost" argument to configure.
It is a pain to have to distribute a boost library at the same
time as the Urbi kernel. This argument makes the static library
versions to be used over the dynamic ones.

The resulting executable + libraries will be smaller as only
the required parts will be embedded. For example, if
boost::thread is linked with but is not required, it will add
zero code to the executable, while the whole libboost-thread-mt.so
would have been required for the kernel to run.

This feature will be sent upstream for integration.