Do not try to join a thread which has not been started.
In pthreads mode, the thread structure is available only when the thread
has been started. This causes a memory access error when trying to terminate
a thread whose structure had not been created.

With this patch, the test suite is clean on x86_64-linux with pthreads,
except for the stack exhaustion check which does not work with pthreads.

	* src/scheduler/pthread-coro.hh (started_): New.
	* src/scheduler/pthread-coro.cc (Coro, coroutine_start, coroutine_free):
	Use it.