More BOOST_CHECK.
I have failures under "make check" that do not happen with ./fd-stream
directly :(  This already happened in the past, because of netcat which
was not able to use large fd numbers (not in the sense of "many", but in
the sense of ">= 16".  Because GNU Make consumes FDs.  I hope this is not
the case here, but this test aborts on my machine.

	* tests/libport/fd-stream.cc: More checking.