Fix check problems on Windows.
The problem is that wine-ld-library was called for each test, each
time erasing and recreating wine/.  Obviously this is wrong with -j.
What makes it harder to fix is that then we no longer have the
certainty that a libtool executable wrapper is available to get the
list of directories holding the libraries we look for.

	* build-aux/wine-ld-library-path.as ($dirs): Remove, replaced by...
        ($windirs, windir): these.
        Don't discard nonexisting dirs.
        (create_wine_directory): No longer define windir, just use it.
	(--ldflags): New.
	* tests/libport/Makefile.am (WINEPREFIX_ENVIRONMENT): Use :=,
        not =, so that it is executed only once.