Simplify the version handling.
Maintaining it is really a PITA, so I decided to simplify: now if you want
to depend on version information, depend on $(top_srcdir)/.version.  This
file is "lazy" (via time stamps), but the others are not.

The target "update-version" updates it (if needed).  This should be
run by the bf.

	* build-aux/package-version.mk: move-if-change is no longer used.
	(VERSION): Be more robust.
	Use a stamp file to create .version.
	Ship both.
	(update-version, debug): New.
	* build-aux/revision.mk (REVISIONFLAGS): New.
	No longer maintain a stamp file.
	* GNUmakefile.in: Remove, we now work in Makefile.am.
	* Makefile.am (package-version.mk): Load it.
	* build-aux/package-version.m4 (GNUmakefile): Forget about it.