Sched: forward termination to the oldest ancester.
When terminating a job, one must beware not to kill the job iterating over
the jobs to kill, as it would immediately cease to do its job.  The code
was paying attention to the job itself, but did not consider the case
of killing an ancester of the current job, which would result in immediate
destruction.

Fix 2.x/eval.chk.

	* include/sched/job.hh, include/sched/job.hxx, lib/sched/job.cc
	(ancester, ancester_of): New.
	(run): Propagate TerminateException to parent.
	* lib/sched/scheduler.cc (killall_jobs): Kill the ancester of the
	current job (and its descendants) last.