libport.semaphore: don't provide sem_getvalue.
This function is not portable (not available at all on OS X) and it's
emulation is more complex than what I did, since, as pointed out by
Quentin Hocquet, this patch was subject to race conditions.

It turns out that changes by Matthieu Nottale on GnetClient made this
feature useless.

And people seem to agree that this function should not be used.

Basically, this reverts dfb3c619fcc7d55ca3f5623b6b5f8543755eceec.

	* include/libport/semaphore.hh, lib/libport/semaphore.cc (value_)
	(operator int): Remove.
	* tests/libport/semaphore.cc: Adjust.