read-stdin: retarget to use FD even on Windows.
Our whole API is using the POSIX interface to files (fopen and so forth).
It seems, to the best of my knowledge, that we cannot mix Windows API
(HANDLE, ReadFile etc.) and POSIX files (int, read).  So it seems saner
to also use the POSIX interface to read stdin, as a regular POSIX file.

	* include/libport/read-stdin.hh (read_fd): New.
	* lib/libport/read-stdin.cc (read_fd): New.
	(ConsumerData): Retarget to fds.
	(read_stdin): Adjust to use it.