cli mess.
The interface for CLI functions used to be "complain and exit".  But
since some of these functions were used in programs that became
libraries, the behavior became inadequate, and instead of exiting the
functions started to throw.

But this ruined all the other places these functions were used.  Start
addressing this.  Lots remains to be done.

	* include/libport/cli.hh, lib/libport/cli.cc: Use a sanitized
	interface.