libportify: revamp the relicense inner loop.
It's too hard, and too fragile, to try to match the whole comment.
One nasty issue is that the RE is too easily too greedy, and it eats
comments that follow.  So instead of making a complex RE that matches
comments, match all the C comments, and then the C++ comments, and on
their content, apply the relicensing regexps.

	* bin/libportify: Do all that.
	* tests/bin/libportify-check.m4sh: Remove useless check.
	* tests/bin/libportify/sample-2.exp,
	* tests/bin/libportify/sample-2.lpc: Fix.