ltdl: set the error to file-not-found when LoadLibrary failed on missing file.
Currently we have traces like these:

[XLTDL] loading urbi/all
[XLTDL]   try_dlopen (work\urbi\all, .la)
[XLTDL]   try_dlopen (work\urbi\all, .dll)
[XLTDL]     tryall_dlopen (work\urbi\all.dll, (ALL))
[XLTDL]       lt_preopen->module_open (work\urbi\all.dll)
[XLTDL]         fail: file not found
[XLTDL]       lt_loadlibrary->module_open (work\urbi\all.dll)
[XLTDL]     LoadLibrary (work\urbi\all.dll)
[XLTDL]     LoadLibrary (work\urbi\all.dll): failed with error 126: Module not found
[XLTDL]         fail: can't open the module
[XLTDL] loading work\urbi\all: fail
[XLTDL]   try_dlopen (lib\urbi\all, .la)
[XLTDL] loading lib\urbi\all: fail

where one can see that it tries to load all.la first in `work', then it
tries all.dll in `work', both fail.  So it tries the next component of the
file search path: `lib'.  But there, it only tries `all.la', it does not
even try `all.dll', which would have succeeded though.  This is because
the first error found during path traversal (here, Module not found, which
sets the Libtool error to CANNOT_OPEN) is preserved if the file exists, so
that we use it to report the error when everything failed.  So it does
matter to report when the file does not exist, i.e., to set the error to
FILE_NOT_FOUND, instead of CANNOT_OPEN.

	* libltdl/loaders/loadlibrary.c (vm_open): here.