xltdl: work around an issue we introduced in ltdl.
This bug was introduced long ago, but it's been uncovered recently
only because we now load the modules without extension, while before
we passed the extension.  Without extension, ltdl first tries *.la,
then the native extension (e.g., *.so).  But if there was an error
after trying the *.la, it does not even try the *.so.

So the loading of the first module without extension succeeds, but
sets the error flag to FILE_NOT_FOUND (because the first attempts did
not find the file).  Then the second load of a module without
extension fails because of the stray error flag.

FIXME: we need a clean solution in ltdl itself.

	* lib/libport/xltdl.cc (open): Clear the error flag.