LIBPORT_BUFSIZ.
It once happened that Shell parser was null in the Kernel.  The
scenario is that an remote UObject in binary mode sends a message that
is too big to fit into the stream buffer, and as a result the message
was incomplete.  This caused processSerializedMessages to end in an
unexpected way, and control returned to the text processing loop in an
unexpected way.  So we returned in the regular processing flow,
although there is no parser around.

To work around this, increase the buffer size (from 1k to 8k on OS X).

	* include/libport/cstdio (LIBPORT_BUFSIZ): New.
	* include/libport/dlfcn.h,
	* include/libport/io-stream.hh,
	* include/libport/iostream,
	* lib/libport/file-system.cc,
	* lib/libport/io-stream.cc,
	* lib/libport/read-stdin.cc,
	* lib/libport/unistd.cc,
	* tests/libport/fd-stream.cc:
	Use it.