check: work with in-place builds.
Issue #5111.

Our tests have names like the corresponding headers.  For instance
include/lipbort/cstdlib is tested by tests/libport/cstdlib.  We used
to pass -I tests, which is fine with VPATH builds, but with in-place
builds, it means that sometimes <libport/cstdlib> would pick up the
compiled test instead of the header.

	* tests/libport/local.mk (AM_CPPFLAGS): Do not pass -Itests.

	* tests/libport/asio.cc, tests/libport/backtrace.cc,
	* tests/libport/cli.cc, tests/libport/hash.cc,
	* tests/libport/input-arguments.cc, tests/libport/option-parser.cc,
	* tests/libport/path.cc, tests/libport/pid-file.cc,
	* tests/libport/read-stdin.cc, tests/libport/thread-pool.cc,
	* tests/libport/time.cc, tests/libport/xltdl.cc,
	* tests/sched/sched-except.cc, tests/sched/sched.cc:
	Adjust.